#57 Change headers type in response and request
